pyproject.toml
[project]
name = "apr-agent"
version = "0.4.0"
description = "Agent-driven APR trajectory producer for Defects4J"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.6",
    "typer>=0.12",
    "rich>=13.0",
    "pyyaml>=6.0",
    "openai>=1.40",          # DashScope compatible-mode endpoint
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
apr-agent = "apr_agent.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/apr_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "defects4j: requires `defects4j` on PATH — skipped otherwise",
    "slow: long-running integration test (checkout/compile/test a real bug)",
    "needs_api_key: requires DASHSCOPE_API_KEY / QWEN_API_KEY in env",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]
